[BACK-END REFACTOR] MOved custom resolvers definition from relations.js to mutations.js
